fixup includes in headers, started as just config, but led to more
The changes in config required changes in ai/contexts which
cascaded, also many places were throwing game::game_error without
including its definition except from config.

+++ src/preferences_display.cpp
@@ -168,7 +168,7 @@ bool set_resolution(CVideo& video
 {
 	SDL_Rect rect;
 	SDL_GetClipRect(video.getSurface(), &rect);
-	if(rect.w == width && rect.h == height) {
+	if(static_cast<unsigned int> (rect.w) == width && static_cast<unsigned int>(rect.h) == height) {
 		return true;
 	}
